Handle async-graphql issue #1365: Produce compile time error when using integer types bigger than i32 (#52)
Source:
https://redirect.github.com/async-graphql/async-graphql/issues/1365

The target still implements the GraphQL Int scalar for i64, u64, usize,
and isize, so using wider Rust integer types remains accepted at compile
time.

Commit: dcd70faa2452bfa35734f957d51b8aea1cc63e2b

+++ src/extensions/analyzer.rs
@@ -83,9 +83,9 @@ mod tests {
             MyObj
         }
 
-        #[graphql(complexity = "count * child_complexity")]
-        async fn objs(&self, count: usize) -> Vec<MyObj> {
-            vec![MyObj; count]
+        #[graphql(complexity = "count as usize * child_complexity")]
+        async fn objs(&self, count: u16) -> Vec<MyObj> {
+            vec![MyObj; count as usize]
         }
     }
